+++ admin/tabs/activity/tracking/ActivityTrackingModel.ts
@@ -429,7 +429,7 @@ export class ActivityTrackingModel extends HoistModel implements ActivityDetailP
             exportOptions: {filename: exportFilename('activity-summary')},
             emptyText: 'No activity reported...',
             sortBy: ['cubeLabel'],
-            expandToLevel: 1,
+            expandLevel: 1,
             levelLabels: () => ['Total', ...this.groupingChooserModel.valueDisplayNames],
             columns: [
                 {