+++ desktop/cmp/leftrightchooser/LeftRightChooserModel.ts
@@ -189,7 +189,7 @@ export class LeftRightChooserModel extends HoistModel {
             onRowDoubleClicked: e => this.onRowDoubleClicked(e),
             columns: [leftTextCol, groupCol],
             contextMenu: false,
-            expandToLevel: leftGroupingExpanded ? 1 : 0,
+            expandLevel: leftGroupingExpanded ? 1 : 0,
             xhImpl: true
         });
 
@@ -201,7 +201,7 @@ export class LeftRightChooserModel extends HoistModel {
             onRowDoubleClicked: e => this.onRowDoubleClicked(e),
             columns: [rightTextCol, groupCol],
             contextMenu: false,
-            expandToLevel: rightGroupingExpanded ? 1 : 0,
+            expandLevel: rightGroupingExpanded ? 1 : 0,
             xhImpl: true
         });